backend/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.1.0
bcrypt==3.2.2
beautifulsoup4==4.12.3
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cryptography==44.0.0
ecdsa==0.19.0
fastapi==0.109.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
iniconfig==2.0.0
multidict==6.1.0
packaging==24.2
passlib==1.7.4
pluggy==1.5.0
propcache==0.2.1
psycopg2-binary==2.9.9
pyasn1==0.6.1
pycparser==2.22
pydantic==2.6.1
pydantic-settings==2.1.0
pydantic_core==2.16.2
pytest==8.3.4
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.6
requests==2.31.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.27
starlette==0.36.3
tenacity==9.0.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.27.1
yarl==1.18.3
